Add special-purpose IdSpan caching codec for messaging and split into separate files
